requirements.txt
torch
torchvision
numpy
opencv-python
pyyaml
tqdm
requests
scikit-image
Pillow